% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rhtmlSankeyTree.R
\name{SankeyTree}
\alias{SankeyTree}
\title{Interactive d3.js Sankey Tree Visualization}
\usage{
SankeyTree(data = NULL, name = "name", id = "id", value = "size",
  childrenName = "children", treeColors = TRUE, maxLabelLength = NULL,
  nodeHeight = NULL, tooltip = NULL, colorLegend = FALSE,
  categoryLegend = FALSE, terminalDescription = FALSE, width = NULL,
  height = NULL)
}
\description{
Create interactive d3.js visualizations of hierarchical data combining
Sankey diagrams with tree layout.
}

