% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhtmlSankeyTree.R
\name{SankeyTree}
\alias{SankeyTree}
\title{Interactive d3.js Sankey Tree Visualization}
\usage{
SankeyTree(data = NULL, name = "name", id = "id", value = "size",
  childrenName = "children", maxLabelLength = 25, nodeHeight = 100,
  tooltip = NULL, treeColors = FALSE, categoryLegend = FALSE,
  terminalDescription = FALSE, numeric.distribution = TRUE,
  footer = NULL, footer.font.size = 11,
  footer.font.family = "sans-serif", footer.font.color = "#000000",
  subtitle = NULL, subtitle.font.size = 18,
  subtitle.font.family = "sans-serif", subtitle.font.color = "#000000",
  title = NULL, title.font.size = 24,
  title.font.family = "sans-serif", title.font.color = "#000000",
  width = NULL, height = NULL)
}
\arguments{
\item{data}{Input. Should be a list with fields specified below.}

\item{name}{The variable name of the field that specifies node name in \code{data}. Defaults to "name".}

\item{id}{Should not change this.}

\item{value}{The variable name of the field that specifies counts in \code{data}. Defaults to "size".}

\item{childrenName}{The variable name of the field that specifies children in \code{data}. Defaults to "children".}

\item{maxLabelLength}{Integer. Line wrapping will occur when the label of the node has width more than this value * 8 (pixels per char). Defaults to 25.}

\item{nodeHeight}{Height of each node in pixels. Defaults to 100 pixels.}

\item{tooltip}{Tooltip to attach to each node.}

\item{terminalDescription}{Logical. Add description for terminal nodes? Defaults to FALSE.}

\item{numeric.distribution}{Logical. Show numerical distribution for each node? If regression tree, then a histogram is plotted. If classification tree, then a horizontal bar plot is drawn. Defaults to TRUE.}

\item{footer}{character. Sets the footer of the chart, defaults to NULL. The footer is left-aligned.}

\item{footer.font.size}{integer. Font size of the chart footer.font.size, defaults to 11 pixcels.}

\item{footer.font.family}{character. Font family of the chart footer.font.family, defaults to "sans-serif".}

\item{footer.font.color}{An RGB character to set the color of the chart footer.font.color. Defaults to "#000000".}

\item{subtitle}{character. Sets the subtitle of the chart, defaults to NULL. The subtitle is centred.}

\item{subtitle.font.size}{integer. Font size of the chart subtitle, defaults to 18 pixcels.}

\item{subtitle.font.family}{character. Font family of the chart subtitle, defaults to "sans-serif".}

\item{subtitle.font.color}{An RGB character to set the color of the chart subtitle. Defaults to "#000000".}

\item{title}{character. Sets the title of the chart, defaults to NULL. The title is centred.}

\item{title.font.size}{integer. Font size of the chart title, defaults to 24 pixcels.}

\item{title.font.family}{character. Font family of the chart title, defaults to "sans-serif".}

\item{title.font.color}{An RGB character to set the color of the chart title. Defaults to "#000000".}
}
\description{
Create interactive d3.js visualizations of hierarchical data combining
Sankey diagrams with tree layout.
}
